appointment_availability/requirements.txt
Flask==2.3.3
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
